backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
google-genai
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-multipart
jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pypdf
python-multipart

# AI & Agentic Workflow
langchain
langgraph
langchain-community
langchain-google-genai
duckduckgo-search
transformers
peft
accelerate
torch
